# core_game_openapi.json
{"openapi":"3.1.0","info":{"title":"Core Game Engine","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/xp/calculate":{"post":{"summary":"Calculate Xp","operationId":"calculate_xp_api_xp_calculate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCompletion"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/XPResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/{uid}":{"get":{"summary":"Get User","operationId":"get_user_api_user__uid__get","parameters":[{"name":"uid","in":"path","required":true,"schema":{"type":"string","title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TaskCompletion":{"properties":{"task_id":{"type":"string","title":"Task Id"},"uid":{"type":"string","title":"Uid"},"difficulty":{"type":"integer","title":"Difficulty"},"mood_coefficient":{"type":"number","title":"Mood Coefficient","default":1.0},"adhd_assist":{"type":"number","title":"Adhd Assist","default":1.0}},"type":"object","required":["task_id","uid","difficulty"],"title":"TaskCompletion"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"XPResponse":{"properties":{"xp_earned":{"type":"integer","title":"Xp Earned"},"new_total_xp":{"type":"integer","title":"New Total Xp"},"level_up":{"type":"boolean","title":"Level Up"},"new_level":{"type":"integer","title":"New Level"}},"type":"object","required":["xp_earned","new_total_xp","level_up","new_level"],"title":"XPResponse"}}}}